fix(codegen): unsigned integer operations used signed semantics
Five bugs where unsigned types (u8/u32/u64) silently fell through to
signed operations:

1. Comparison (<, >, <=, >=) used slt/sgt/sle/sge instead of ult/ugt/ule/uge
2. Division used DivSIOp instead of DivUIOp
3. Modulo used RemSIOp instead of RemUIOp
4. ECast widening used ExtSIOp instead of ExtUIOp for unsigned sources
5. match PatInt patterns truncated to int32_t — fix lir.hpp + sema.cpp

All bugs confirmed with .logos repros before fixing. 4 regression
tests added; suite: 271 pass / 362 total.

+++ include/logos/compiler/lir.hpp
@@ -41,7 +41,7 @@ using LBlockPtr = std::unique_ptr<LBlock>;
 // ── Patterns (for match arms) ─────────────────────────────────────────────
 
 struct PatVariant { std::string enum_name; std::string variant; int32_t disc; };
-struct PatInt     { int32_t value; };
+struct PatInt     { int64_t value; };
 struct PatBool    { bool value; };
 struct PatWild    { std::string name; };   // _ or named wildcard (name may be "_")
